[refactor]Home: separate UI from logic
